refactor: 640 - add icons aws and add in type Course - iconFooter
