+++ dev-data/courses.data.ts
@@ -1,12 +1,10 @@
 import { ROUTES } from '@/core/const';
 import type { Course } from '@/entities/course';
+import aws from '@/shared/assets/icons/amazon-web-services-logo-wine-256.svg';
+import footerIcon from '@/shared/assets/icons/footer/amazon-web-services-logo-footer-256.svg';
 import angular from '@/shared/assets/icons/angular-256.svg';
 import angularSecondary from '@/shared/assets/icons/angular-black-256.svg';
-import awsSecondary from '@/shared/assets/icons/aws-secondary.webp';
-import aws from '@/shared/assets/icons/aws.svg';
 import angularSmall from '@/shared/assets/icons/footer/angular-256.svg';
-import awsDevSmall from '@/shared/assets/icons/footer/aws-dev.webp';
-import awsFundSmall from '@/shared/assets/icons/footer/aws-fundamentals.webp';
 import htmlSmall from '@/shared/assets/icons/footer/html5-256.svg';
 import jsSmall from '@/shared/assets/icons/footer/javascript-256.svg';
 import nodejsSmall from '@/shared/assets/icons/footer/node-js-256-small.svg';
@@ -29,6 +27,7 @@ export const courses: Course[] = [
     altTitle: 'JavaScript / Front-end',
     iconSrc: javascript,
     iconSmall: htmlSmall,
+    iconFooter: htmlSmall,
     secondaryIcon: jsSecondary,
     startDate: TO_BE_DETERMINED,
     registrationEndDate: TO_BE_DETERMINED,
@@ -49,6 +48,7 @@ export const courses: Course[] = [
     altTitle: 'JavaScript / Front-end',
     iconSrc: javascript,
     iconSmall: jsSmall,
+    iconFooter: jsSmall,
     secondaryIcon: jsSecondary,
     startDate: TO_BE_DETERMINED,
     registrationEndDate: TO_BE_DETERMINED,
@@ -69,6 +69,7 @@ export const courses: Course[] = [
     altTitle: 'JavaScript / Front-end',
     iconSrc: javascript,
     iconSmall: jsSmall,
+    iconFooter: jsSmall,
     secondaryIcon: jsSecondary,
     startDate: TO_BE_DETERMINED,
     registrationEndDate: TO_BE_DETERMINED,
@@ -88,6 +89,7 @@ export const courses: Course[] = [
     descriptionUrl: COURSE_LINKS.REACT,
     iconSrc: react,
     iconSmall: reactSmall,
+    iconFooter: reactSmall,
     secondaryIcon: reactSecondary,
     startDate: TO_BE_DETERMINED,
     registrationEndDate: TO_BE_DETERMINED,
@@ -107,6 +109,7 @@ export const courses: Course[] = [
     descriptionUrl: COURSE_LINKS.ANGULAR,
     iconSrc: angular,
     iconSmall: angularSmall,
+    iconFooter: angularSmall,
     secondaryIcon: angularSecondary,
     startDate: TO_BE_DETERMINED,
     registrationEndDate: TO_BE_DETERMINED,
@@ -126,6 +129,7 @@ export const courses: Course[] = [
     descriptionUrl: COURSE_LINKS.NODE,
     iconSrc: nodejs,
     iconSmall: nodejsSmall,
+    iconFooter: nodejsSmall,
     secondaryIcon: nodejsSecondary,
     startDate: TO_BE_DETERMINED,
     registrationEndDate: TO_BE_DETERMINED,
@@ -144,8 +148,9 @@ export const courses: Course[] = [
     subTitle: null,
     descriptionUrl: COURSE_LINKS.AWS_FUNDAMENTALS,
     iconSrc: aws,
-    iconSmall: awsFundSmall,
-    secondaryIcon: awsSecondary,
+    iconSmall: aws,
+    iconFooter: footerIcon,
+    secondaryIcon: aws,
     startDate: TO_BE_DETERMINED,
     registrationEndDate: TO_BE_DETERMINED,
     language: 'en',
@@ -164,8 +169,9 @@ export const courses: Course[] = [
     subTitle: null,
     descriptionUrl: COURSE_LINKS.AWS_CLOUD_DEVELOPER,
     iconSrc: aws,
-    iconSmall: awsDevSmall,
-    secondaryIcon: awsSecondary,
+    iconSmall: aws,
+    secondaryIcon: aws,
+    iconFooter: footerIcon,
     startDate: TO_BE_DETERMINED,
     registrationEndDate: TO_BE_DETERMINED,
     language: 'en',
@@ -183,8 +189,9 @@ export const courses: Course[] = [
     subTitle: null,
     descriptionUrl: COURSE_LINKS.AWS_DEVOPS,
     iconSrc: aws,
-    iconSmall: awsDevSmall,
-    secondaryIcon: awsSecondary,
+    iconSmall: aws,
+    secondaryIcon: aws,
+    iconFooter: footerIcon,
     startDate: TO_BE_DETERMINED,
     registrationEndDate: TO_BE_DETERMINED,
     language: 'en',
